refactor entity availability by implementing specific entity mixins
